Feature/user create endpoint

+++ BackEnd/src/main/java/com/rmit/sept/assignment/initial/service/HoursService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.DayOfWeek;
+import java.time.LocalTime;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -29,22 +30,16 @@ public class HoursService {
         if (workerId == null) return null;  // must have an id value
         Optional<Worker> existingWorker = workerRepository.findById(workerId);
         if (!existingWorker.isPresent()) return null;  // worker must already exist in the system
-        Date start = hours.getStart();
-        Date end = hours.getEnd();
+        LocalTime start = hours.getStart();
+        LocalTime end = hours.getEnd();
         if (!(start == null || end == null)) {
-            Calendar calStart = Calendar.getInstance();
-            Calendar calEnd = Calendar.getInstance();
-            calStart.setTime(start);
-            calEnd.setTime(end);
-            boolean sameDay = (calEnd.get(Calendar.DAY_OF_YEAR) == calStart.get(Calendar.DAY_OF_YEAR)) &&
-                    (calEnd.get(Calendar.YEAR) == calStart.get(Calendar.YEAR));
-            if (!sameDay || calStart.after(calEnd)) return null;
+            if (start.compareTo(end) >= 0) return null;
         }
         return hoursRepository.save(hours);
     }
 
     public boolean deleteHours(Hours hours) {
-        if (hours.getId() == null) return false;
+        if (hours == null || hours.getId() == null) return false;
         Optional<Hours> h1 = hoursRepository.findById(hours.getId());
         if (h1.isPresent()) {
             hoursRepository.deleteById(hours.getId());
@@ -59,7 +54,7 @@ public class HoursService {
             Hours.HoursPK hoursPK = new Hours.HoursPK(worker.get(), dayOfWeek);
             Optional<Hours> hours = hoursRepository.findById(hoursPK);
             if (hours.isPresent()) {
-                hoursRepository.deleteById(hoursPK);
+                hoursRepository.delete(hours.get());
                 return !hoursRepository.findById(hoursPK).isPresent();  // return true if hours have been removed
             }
         }
@@ -79,44 +74,4 @@ public class HoursService {
         if (worker == null) return null;
         return hoursRepository.findById_Worker(worker);
     }
-//
-//    public Hours createHours(Hours hours) {
-//        System.out.println(hours);
-//        Hours h1 = hoursRepository.save(hours);
-//        System.out.println(h1);
-//        return h1;
-//    }
-//
-//    public Hours findById(Worker worker, int dayOfWeek) {
-//        if (worker == null || dayOfWeek <= -1 || dayOfWeek > 6) return null;
-//        return hoursRepository.findById(new HoursPK(worker.getId(), dayOfWeek)).orElse(null);
-//    }
-//
-//    public List<Hours> findByWorker(Worker worker) {
-//        if (worker == null) return null;
-//        return hoursRepository.findAllByWorker(worker);
-//    }
-
-//
-//    public Hours createHours(Hours hours) {
-//        if (hours.getWorker() == null) return null;
-//        Long workerId = hours.getWorker().getId();
-//        System.out.println("WORKERID " + workerId);
-//        Optional<Worker> worker = workerRepository.findById(workerId);
-//        if (worker.isPresent()) {
-//            System.out.println("PRESENT");
-//            System.out.println(hours.getStart() + " " + hours.getEnd());
-//            Hours h2 =  hoursRepository.save(hours);
-//            System.out.println(h2);
-//            return h2;
-////            if (!hoursRepository.findByWorkerAndAndDayOfWeek(worker.get().getId(), hours.getDayOfWeek()).isPresent()) {
-////            }
-//        }
-//        return null;
-//    }
-//
-//    public Hours findById(Long id) {
-//        Optional<Hours> hours =  hoursRepository.findById(id);
-//        return hours.orElse(null);
-//    }
 }
